team controller, tests un peu partout

+++ api/src/controllers/user.controller.ts
@@ -1,11 +1,7 @@
 import { Article, PrismaClient, Role } from "@prisma/client";
 import { Request, Response } from "express";
 import { hashPassword } from "../utils/security";
-
-export interface ArticleInterface {
-  title: string;
-  body: string;
-}
+import { ArticleInterface } from "./article.controller";
 
 export interface UserInterface {
   username: string;
@@ -35,6 +31,11 @@ export const getAllUsers = async (req: Request, res: Response) => {
 export const getUserById = async (req: Request, res: Response) => {
   try {
     const userId = req.params.id;
+    const userIdValidation = !userId || isNaN(Number(userId));
+    if (userIdValidation) {
+      res.status(404).json({ message: "ID invalide : ", userId });
+      return;
+    }
 
     const user = await userClient.findUnique({
       where: {
@@ -52,17 +53,25 @@ export const getUserById = async (req: Request, res: Response) => {
 // createUser
 export const createUser = async (req: Request, res: Response) => {
   try {
-    const { username, password, role, articles }: UserInterface = req.body;
-    const hashedPassword = hashPassword(password);
+    const { username, password, role, articles }: Partial<UserInterface> =
+      req.body;
+
+    const userBodyValidation = !username || !password || !role;
+    if (userBodyValidation) {
+      res.status(404).json({ message: "Veuillez compléter tous les champs" });
+      return;
+    }
+
+    const pw = password as string;
+    const psd = username as string;
+    const hashedPassword = hashPassword(pw);
 
     const user = await userClient.create({
       data: {
-        username,
+        username: psd,
         password: hashedPassword,
         role: role || "USER",
-        articles: {
-          create: articles,
-        },
+        articles: articles?.length ? { create: articles } : undefined,
       },
     });
 
@@ -76,29 +85,35 @@ export const createUser = async (req: Request, res: Response) => {
 export const updateUser = async (req: Request, res: Response) => {
   try {
     const userId = req.params.id;
+    const userIdValidation = !userId || isNaN(Number(userId));
+    if (userIdValidation) {
+      res.status(404).json({ message: "ID invalide : ", userId });
+      return;
+    }
 
-    const { username, password, role, articles }: UserInterface = req.body;
-    const hashedPassword = hashPassword(password);
+    const { username, password, role, articles }: Partial<UserInterface> =
+      req.body;
 
-    const articlesOperations = articles?.map((article: ArticleInterface) => ({
-      title: article.title,
-      body: article.body,
-    }));
+    const updateData: any = {};
+    if (username) updateData.username = username;
+    if (password) updateData.password = hashPassword(password);
+    if (role) updateData.role = role;
+    if (articles) {
+      updateData.articles = {
+        create: articles.map((article) => ({
+          title: article.title,
+          body: article.body,
+        })),
+      };
+    }
 
     const user = await userClient.update({
       where: { id: Number(userId) },
-      data: {
-        username,
-        password: hashedPassword,
-        role: role || "USER",
-        articles: {
-          create: articlesOperations,
-        },
-      },
+      data: updateData,
       include: { articles: true },
     });
 
-    res.status(201).json({ data: user });
+    res.status(200).json({ data: user });
   } catch (e) {
     console.log(e);
   }
@@ -108,12 +123,17 @@ export const updateUser = async (req: Request, res: Response) => {
 export const deleteUser = async (req: Request, res: Response) => {
   try {
     const userId = req.params.id;
+    const userIdValidation = !userId || isNaN(Number(userId));
+    if (userIdValidation) {
+      res.status(400).json({ message: "ID invalide : ", userId });
+      return;
+    }
 
     const user = await userClient.delete({
       where: { id: Number(userId) },
     });
 
-    res.status(200).json({ data: user });
+    res.status(204).json({ data: user });
   } catch (e) {
     console.log(e);
   }
